Clear the twelve clippy findings batch 3 left
None of the eight workers could run clippy: for most of the afternoon the
crate did not compile, because some sibling was always mid-edit, and
`cargo clippy --fix` has no path filter and would have rewritten files it
did not own. So the one `--fix` pass the recipe asks for per family was
deferred to integration, which is this commit.

Ten are what a `Live<T>` conversion always leaves -- `explicit_auto_deref`
where the handle already derefs, `unnecessary_cast` where a length stopped
being a `c_int`, one `redundant_closure`, one `collapsible_if`.

Two were worth reading:

* **`ex_docmd/quit.rs`'s `before_quit_autocmds` was a *safe* `pub fn` taking
  `*mut win_T` and dereferencing it** (`clippy::not_unsafe_ptr_arg_deref`,
  deny-level). Its three callers all sit in the same file, so it is
  `pub(crate) unsafe fn` now with the obligation written down: `wp` must be
  live on entry and need not survive the call, which is exactly what
  `quit_was_cancelled` is there to re-check.
* `insexpand/matchlist.rs` tested `first.is_none()` and then `expect`ed it in
  the next arm. The `if`/`else if`/`else` is a `match` on `first_match()`
  now, with the guard on the `Some` arm, so the value is bound rather than
  re-asserted.

`metrics/clippy.json` is unchanged at 19 warnings in 16 files.

+++ crates/nvim/src/ex_docmd/filename.rs
@@ -306,13 +306,13 @@ pub(crate) fn repl_cmdline(
         )
     };
     let tail = offset + len;
-    unsafe { strcpy(new_cmdline.add(tail as usize), src.add(srclen)) };
-    let resume = unsafe { new_cmdline.add(tail as usize) };
+    unsafe { strcpy(new_cmdline.add(tail), src.add(srclen)) };
+    let resume = unsafe { new_cmdline.add(tail) };
 
     if !ea.nextcmd.is_null() {
         let after = strlen(new_cmdline) + 1;
-        unsafe { strcpy(new_cmdline.add(after as usize), ea.nextcmd) };
-        ea.nextcmd = unsafe { new_cmdline.add(after as usize) };
+        unsafe { strcpy(new_cmdline.add(after), ea.nextcmd) };
+        ea.nextcmd = unsafe { new_cmdline.add(after) };
     }
     ea.cmd = unsafe { new_cmdline.offset(ea.cmd.offset_from(*cmdlinep)) };
     ea.arg = unsafe { new_cmdline.offset(ea.arg.offset_from(*cmdlinep)) };
